vanilla qwik-template
