--- packages/create-vite/template-qwik-ts/_gitignore
@@ -1,38 +1,0 @@
-# Build
-/dist
-/lib
-/lib-types
-/server
-
-# Development
-node_modules
-
-# Cache
-.cache
-.mf
-.vscode
-.rollup.cache
-tsconfig.tsbuildinfo
-
-# Logs
-logs
-*.log
-npm-debug.log*
-yarn-debug.log*
-yarn-error.log*
-pnpm-debug.log*
-lerna-debug.log*
-
-# Editor
-!.vscode/extensions.json
-.idea
-.DS_Store
-*.suo
-*.ntvs*
-*.njsproj
-*.sln
-*.sw?
-
-# Yarn
-.yarn/*
-!.yarn/releases
