+++ packages/create-vite/template-qwik-ts/package.json
@@ -1,37 +1,18 @@
 {
-  "name": "my-qwik-library-name",
-  "version": "0.0.1",
-  "description": "Create a reusable Qwik component library",
-  "engines": {
-    "node": ">=15.0.0"
-  },
-  "private": false,
+  "name": "vite-qwik",
+  "private": true,
+  "version": "0.0.0",
   "type": "module",
   "scripts": {
-    "build": "qwik build",
-    "build.lib": "vite build --mode lib",
-    "build.types": "tsc --emitDeclarationOnly",
-    "dev": "vite --mode ssr",
-    "dev.debug": "node --inspect-brk ./node_modules/vite/bin/vite.js --mode ssr --force",
-    "fmt": "prettier --write .",
-    "fmt.check": "prettier --check .",
-    "lint": "eslint \"src/**/*.ts*\"",
-    "release": "np",
-    "start": "vite --open --mode ssr",
-    "qwik": "qwik"
+    "dev": "vite",
+    "build": "tsc && node --inspect-brk node_modules/vite/bin/vite.js build",
+    "preview": "node node_modules/vite/bin/vite.js preview"
   },
   "devDependencies": {
-    "@builder.io/qwik": "1.1.5",
-    "@types/eslint": "8.40.0",
-    "@types/node": "^20.2.5",
-    "@typescript-eslint/eslint-plugin": "5.59.8",
-    "@typescript-eslint/parser": "5.59.8",
-    "eslint": "8.41.0",
-    "eslint-plugin-qwik": "latest",
-    "np": "7.6.1",
-    "prettier": "2.8.8",
-    "typescript": "5.0.4",
-    "undici": "5.22.1",
-    "vite": "4.3.9"
+    "typescript": "^5.0.2",
+    "vite": "^4.3.9"
+  },
+  "dependencies": {
+    "@builder.io/qwik": "^1.2.5"
   }
 }
