+++ packages/create-vite/template-qwik-ts/src/app.tsx
@@ -1,0 +1,34 @@
+import { component$, useSignal } from '@builder.io/qwik'
+
+import qwikLogo from './assets/qwik.svg'
+import viteLogo from '/vite.svg'
+import './app.css'
+
+export const App = component$(() => {
+  const count = useSignal(0)
+
+  return (
+    <>
+      <div>
+        <a href="https://vitejs.dev" target="_blank">
+          <img src={viteLogo} class="logo" alt="Vite logo" />
+        </a>
+        <a href="https://qwik.builder.io" target="_blank">
+          <img src={qwikLogo} class="logo qwik" alt="Qwik logo" />
+        </a>
+      </div>
+      <h1>Vite + Qwik</h1>
+      <div class="card">
+        <button onClick$={() => count.value++}>
+          count is {count.value}
+        </button>
+        <p>
+          Edit <code>src/app.tsx</code> and save to test HMR
+        </p>
+      </div>
+      <p class="read-the-docs">
+        Click on the Vite and Qwik logos to learn more
+      </p>
+    </>
+  )
+});
