--- packages/create-vite/template-qwik-ts/src/entry.ssr.tsx
@@ -1,25 +1,0 @@
-/**
- * WHAT IS THIS FILE?
- *
- * SSR entry point, in all cases the application is rendered outside the browser, this
- * entry point will be the common one.
- *
- * - Server (express, cloudflare...)
- * - npm run start
- * - npm run preview
- * - npm run build
- *
- */
-import {
-  renderToStream,
-  type RenderToStreamOptions,
-} from '@builder.io/qwik/server'
-import { manifest } from '@qwik-client-manifest'
-import Root from './root'
-
-export default function (opts: RenderToStreamOptions) {
-  return renderToStream(<Root />, {
-    manifest,
-    ...opts,
-  })
-}
