--- packages/create-vite/template-qwik-ts/src/index.ts
@@ -1,1 +1,0 @@
-export { App } from './components/app/app'
