+++ packages/create-vite/template-qwik-ts/src/main.tsx
@@ -1,0 +1,7 @@
+import '@builder.io/qwik/qwikloader.js'
+
+import { render } from '@builder.io/qwik'
+import { App } from './app.tsx'
+import './index.css'
+
+render(document.getElementById('app') as HTMLElement, <App />)
