+++ packages/create-vite/template-qwik-ts/vite.config.ts
@@ -1,16 +1,9 @@
 import { defineConfig } from 'vite'
-import { qwikVite } from '@builder.io/qwik/optimizer'
+import { qwikVite } from '@builder.io/qwik/optimizer';
 
-export default defineConfig(() => {
-  return {
-    build: {
-      target: 'es2020',
-      lib: {
-        entry: './src/index.ts',
-        formats: ['es', 'cjs'],
-        fileName: (format) => `index.qwik.${format === 'es' ? 'mjs' : 'cjs'}`,
-      },
-    },
-    plugins: [qwikVite()],
-  }
+// https://vitejs.dev/config/
+export default defineConfig({
+  plugins: [qwikVite({
+    csr: true,
+  })],
 })
